# docs/OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sumurai Financial API",
    "description": "Multi-tenant financial aggregation platform with unified Plaid and Teller integration",
    "contact": {
      "name": "Sumurai Support",
      "url": "https://github.com/two-bit-foundry/sumurai"
    },
    "license": {
      "name": "Sustainable Use License v1.0",
      "url": "https://github.com/two-bit-foundry/sumurai/blob/main/LICENSE"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "http://localhost:8080/api",
      "description": "Local via Nginx proxy"
    },
    {
      "url": "/api",
      "description": "Relative API path (use in production)"
    }
  ],
  "paths": {
...
    "/api/billing/checkout": {
      "post": {
        "tags": [
          "Billing"
        ],
        "operationId": "create_billing_checkout",
        "responses": {
          "200": {
            "description": "Paddle checkout created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingCheckoutResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Billing disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Paddle request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_cookie": []
          }
        ]
      }
    },
    "/api/billing/payment-method": {
      "post": {
        "tags": [
          "Billing"
        ],
        "operationId": "create_billing_payment_method",
        "responses": {
          "200": {
            "description": "Payment method checkout transaction created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingCheckoutResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Billing disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Entitlement unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Paddle request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_cookie": []
          }
        ]
      }
    },
    "/api/billing/portal-session": {
      "post": {
        "tags": [
          "Billing"
        ],
        "operationId": "create_billing_portal_session",
        "responses": {
          "200": {
            "description": "Billing portal session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingPortalSessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Billing disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Entitlement unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Paddle request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_cookie": []
          }
        ]
      }
    },
    "/api/billing/status": {
      "get": {
        "tags": [
          "Billing"
        ],
        "operationId": "get_authenticated_billing_status",
        "responses": {
          "200": {
            "description": "Billing status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_cookie": []
          }
        ]
      }
    },
    "/api/billing/trials/start": {
      "post": {
        "tags": [
          "Billing"
        ],
        "operationId": "start_open_trial",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrialStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Open trial start pending webhook fulfillment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrialStartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid trial start",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Billing or trials disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Trial already used",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Paddle request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_cookie": []
          }
        ]
      }
    },
    "/api/billing/webhooks/paddle": {
      "post": {
        "tags": [
          "Billing"
        ],
        "operationId": "handle_paddle_billing_webhook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Webhook accepted"
          },
          "400": {
            "description": "Invalid webhook",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Billing disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "details": {},
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "code": "AUTH_REQUIRED",
          "details": null,
          "error": "UNAUTHORIZED",
          "message": "Authentication required."
        }
      },
...
      "BillingCheckoutResponse": {
        "type": "object",
        "required": [
          "checkout_url",
          "transaction_id"
        ],
        "properties": {
          "checkout_url": {
            "type": "string"
          },
          "transaction_id": {
            "type": "string"
          }
        }
      },
      "BillingPortalSessionResponse": {
        "type": "object",
        "required": [
          "overview_url",
          "subscription_urls"
        ],
        "properties": {
          "overview_url": {
            "type": "string"
          },
          "subscription_urls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "BillingStatusResponse": {
        "type": "object",
        "required": [
          "billing_enabled",
          "trials_enabled",
          "access_status",
          "can_use_own_data",
          "is_demo_mode_active",
          "payment_method_required",
          "billing_portal_available",
          "enabled_financial_providers"
        ],
        "properties": {
          "access_status": {
            "type": "string"
          },
          "billing_enabled": {
            "type": "boolean"
          },
          "billing_portal_available": {
            "type": "boolean"
          },
          "can_use_own_data": {
            "type": "boolean"
          },
          "current_period_ends_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "enabled_financial_providers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "is_demo_mode_active": {
            "type": "boolean"
          },
          "payment_method_required": {
            "type": "boolean"
          },
          "trial_ends_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "trials_enabled": {
            "type": "boolean"
          }
        }
      },
...
      "TrialStartRequest": {
        "type": "object",
        "required": [
          "country_code",
          "postal_code"
        ],
        "properties": {
          "country_code": {
            "type": "string"
          },
          "postal_code": {
            "type": "string"
          }
        }
      },
      "TrialStartResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "auth_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_token",
        "description": "HttpOnly auth cookie issued by the authentication endpoints."
      }
    }
  },
  "security": [
    {
      "auth_cookie": []
    }
  ],
...
}